fix: make the scenario content hash architecture-independent
The first live draft build failed exactly as the safety net intends:
the builder recomputed the bundle hash and rejected the mismatch. Root
cause: hash_field fed usize::to_le_bytes() into the digest — 4 bytes
on wasm32, 8 bytes on x86_64 — so the browser-side wasm produced
different hashes than every native pipeline. Cast to u64 explicitly
(the bytes 64-bit native builds always produced, so no existing image
hashes change) and regenerate the wasm pkg; the committed wasm now
hashes all three repo scenarios byte-identically to intar-image-cli.

+++ crates/intar-image-scenario/src/content_hash.rs
@@ -94,7 +94,10 @@ pub fn normalize_relative_path(path: &str) -> Result<String, ScenarioError> {
 pub fn hash_field(hasher: &mut Sha256, key: &str, value: &[u8]) {
     hasher.update(key.as_bytes());
     hasher.update([0]);
-    hasher.update(value.len().to_le_bytes());
+    // Explicit u64: usize::to_le_bytes() is 4 bytes on wasm32 but 8 on
+    // x86_64, which made the hash architecture-dependent. u64 matches the
+    // bytes every 64-bit native build has always produced.
+    hasher.update((value.len() as u64).to_le_bytes());
     hasher.update(value);
     hasher.update([0xff]);
 }
